♻️ refactor(composable-screen): split elements into separate files
- extract each UIElement type + Zod schema into elements/<Type>Element.ts
- extract each React component into elements/<Type>Element.tsx
- add elements/shared.ts (RenderContext, interpolate) and renderElement.tsx
- types.ts: 443→173 lines, Renderer.tsx: 630→58 lines
